+++ noodles-sam/src/alignment.rs
@@ -1,0 +1,5 @@
+//! Alignment record.
+
+mod record;
+
+pub use self::record::Record;